backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
boto3==1.26.2
botocore==1.29.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.51.0
django-appconf==1.0.5
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-imagekit==4.1.0
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pilkit==2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
